refactor: delete files, rename files
