+++ src/tests/setup/msw/handlers.ts
@@ -1,4 +1,3 @@
 import graph from './handlers/graph.ts';
-import rest from './handlers/rest.ts';
 
-export const handlers = [...graph, ...rest];
+export const handlers = [...graph];
