--- src/tests/setup/msw/handlers/rest.ts
@@ -1,7 +1,0 @@
-import { http, HttpResponse } from 'msw';
-
-const handlers = [
-  http.get('https://swapi.dev/api/people', () => HttpResponse.json({})),
-];
-
-export default handlers;
