+++ src/widgets/graphql-client/graphql-client.test.tsx
@@ -2,7 +2,7 @@ import { screen } from '@testing-library/react';
 import { ReactElement } from 'react';
 import * as firebaseHooks from 'react-firebase-hooks/auth';
 
-import { userMock } from '@/tests/setup/mocks/userMock';
+import { userMock } from '@/tests/setup/mocks/user-mock';
 import { renderWithProviders } from '@/tests/setup/render-router';
 
 import {
